Update platforms list to match new test matrix
We are only interested in supporting the latest releases/LTSes of each
distro, so we need to update the list to match what we test. Also we
added SUSE support, thought we fudge it a bit to claim SLES support
despite testing against OpenSUSE. I won't tell if you don't.